3.1.6. Move IntArrayStack to proper package
