+++ src/main/java/edu/kis/vh/nursery/list/IntArrayStack.java
@@ -1,4 +1,4 @@
-package edu.kis.vh.nursery;
+package edu.kis.vh.nursery.list;
 
 import edu.kis.vh.nursery.list.Stack;
